chore: moving item utils function to independent file
